+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/binder/ViewBinder.kt
@@ -1,0 +1,9 @@
+package me.ibrahimyilmaz.kiel.binder
+
+import android.view.View
+
+abstract class ViewBinder<in T>(private val itemView: View) {
+
+    abstract fun bind(item: T)
+    open fun bind(item: T, payload: List<Any>) = Unit
+}