+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/datasource/DataSource.kt
@@ -1,16 +1,18 @@
 package me.ibrahimyilmaz.kiel.datasource
 
 import androidx.recyclerview.widget.RecyclerView
-import me.ibrahimyilmaz.kiel.binder.ViewHolderBinder
+import me.ibrahimyilmaz.kiel.renderer.Renderer
 import java.lang.ref.WeakReference
 
 abstract class DataSource<T : Any, A : RecyclerView.Adapter<*>>(
-    private val renderers: Map<Class<*>, ViewHolderBinder<T>>
+    renderers: List<Renderer<T>>
 ) {
     private val viewTypeToRendererKeyMap = renderers
         .map {
-            it.value.itemViewType to it.key
+            it.itemViewType to it
         }.toMap()
+
+    internal val rendererMap = renderers.map { it.key to it }.toMap()
 
     internal var recyclerViewAdapter = WeakReference<A>(null)
 
@@ -20,6 +22,6 @@ abstract class DataSource<T : Any, A : RecyclerView.Adapter<*>>(
 
     fun getRendererOf(
         viewType: Int
-    ) = requireNotNull(renderers[viewTypeToRendererKeyMap[viewType]])
+    ) = requireNotNull(viewTypeToRendererKeyMap[viewType])
 }