+++ kiel/src/test/java/me/ibrahimyilmaz/kiel/datasource/RecyclerDataSourceTest.kt
@@ -1,20 +1,22 @@
 package me.ibrahimyilmaz.kiel.datasource
 
 import com.google.common.truth.Truth.assertThat
-import me.ibrahimyilmaz.kiel.binder.LayoutResourceViewHolderBinder
+import me.ibrahimyilmaz.kiel.renderer.Renderer
 import me.ibrahimyilmaz.kiel.datasource.util.TestItem
 import me.ibrahimyilmaz.kiel.datasource.util.TestItemTwo
-import me.ibrahimyilmaz.kiel.datasource.util.TestItemViewHolderBinder
-import me.ibrahimyilmaz.kiel.datasource.util.TestItemViewHolderBinderTwo
+import me.ibrahimyilmaz.kiel.datasource.util.TestItemTwoViewBinder
+import me.ibrahimyilmaz.kiel.datasource.util.TestItemViewBinder
 import org.junit.Before
 import org.junit.Test
 
 class RecyclerDataSourceTest {
 
-    private val viewHolderBinder: LayoutResourceViewHolderBinder<Any> =
-        TestItemViewHolderBinder(1)
-    private val viewHolderBinderTwo: LayoutResourceViewHolderBinder<Any> =
-        TestItemViewHolderBinderTwo(2)
+    private val rendererOne: Renderer<Any> =
+        Renderer(TestItem::class.java, 1, ::TestItemViewBinder)
+
+    private val rendererTwo: Renderer<Any> =
+        Renderer(TestItemTwo::class.java, 2, ::TestItemTwoViewBinder)
+
     private val itemOne = TestItem(1)
     private val itemTwo =
         TestItemTwo("ibra")
@@ -25,9 +27,9 @@ class RecyclerDataSourceTest {
     @Before
     fun setUp() {
         dataSource = RecyclerDataSource(
-            mapOf(
-                TestItem::class.java to viewHolderBinder,
-                TestItemTwo::class.java to viewHolderBinderTwo
+            listOf(
+                rendererOne,
+                rendererTwo
             )
         )
 
@@ -37,10 +39,10 @@ class RecyclerDataSourceTest {
     @Test
     fun `Should return correct renderer for the given type`() {
         // WHEN
-        val actualRenderer = dataSource.getRendererOf(viewHolderBinder.itemViewType)
+        val actualRenderer = dataSource.getRendererOf(rendererOne.itemViewType)
 
         // THEN
-        assertThat(actualRenderer).isEqualTo(viewHolderBinder)
+        assertThat(actualRenderer).isEqualTo(rendererOne)
     }
 
     @Test
@@ -49,7 +51,7 @@ class RecyclerDataSourceTest {
         val actualItemViewType = dataSource.getItemViewType(0)
 
         // THEN
-        assertThat(actualItemViewType).isEqualTo(viewHolderBinder.itemViewType)
+        assertThat(actualItemViewType).isEqualTo(rendererOne.itemViewType)
     }
 
     @Test