+++ kiel/src/test/java/me/ibrahimyilmaz/kiel/datasource/util/TestItemTwoViewBinder.kt
@@ -1,0 +1,14 @@
+package me.ibrahimyilmaz.kiel.datasource.util
+
+import android.view.View
+import me.ibrahimyilmaz.kiel.binder.ViewBinder
+
+class TestItemTwoViewBinder(view: View) : ViewBinder<TestItemTwo>(view) {
+    override fun bind(item: TestItemTwo) {
+        TODO("Not yet implemented")
+    }
+
+    override fun bind(item: TestItemTwo, payload: List<Any>) {
+        TODO("Not yet implemented")
+    }
+}